refactor-structure-project
deleted:    public/vite.svg
modified:   src/App.jsx
deleted:    src/assets/react.svg
new file:   src/components/AppBar/index.jsx
new file:   src/components/ModeSelect/index.jsx
new file:   src/pages/Auth/index.jsx
new file:   src/pages/Boards/BoardBar/index.jsx
new file:   src/pages/Boards/BoardContent/index.jsx
new file:   src/pages/Boards/_id.jsx
new file:   src/pages/Boards/index.jsx
new file:   src/pages/Users/_id.jsx
new file:   src/pages/Users/index.jsx
