+++ src/pages/Boards/BoardBar/index.jsx
@@ -1,0 +1,17 @@
+import Box from '@mui/material/Box'
+
+function index() {
+  return (
+    <Box sx={{
+      backgroundColor: 'primary.dark',
+      width: '100%',
+      height: (theme) => theme.trello.boardBarHeight,
+      display: 'flex',
+      alignItems: 'center'
+    }}>
+      Board Bar
+    </Box>
+  )
+}
+
+export default index
