+++ src/pages/Users/_id.jsx
@@ -1,0 +1,1 @@
+// User detail
