+++ src/pages/Users/index.jsx
@@ -1,0 +1,1 @@
+// Users list
